Organize components by view to which they belong
