+++ interface/frontend/src/components/annotate-and-segment/Annotate.vue
