+++ interface/frontend/src/components/annotate-and-segment/ConcernSlider.vue
