+++ interface/frontend/src/components/annotate-and-segment/Nodule.vue
