+++ interface/frontend/src/components/annotate-and-segment/NoduleList.vue
