+++ interface/frontend/src/components/annotate-and-segment/RadioInput.vue
