+++ interface/frontend/src/components/detect-and-select/Candidate.vue
