+++ interface/frontend/src/components/detect-and-select/CandidateList.vue
