+++ interface/frontend/src/components/open-image/ImageSeries.vue
