+++ interface/frontend/src/components/open-image/TreeView.vue
